+++ task/base/build.js
@@ -4,6 +4,6 @@ var gulp = require('gulp');
 var sequence = require('run-sequence');
 var include = require('require-dir');
 
-gulp.task('build', ['setup'], function(callback) {
-	sequence(Object.keys(include('build/')), callback);
+gulp.task('build', function(callback) {
+	sequence('reset', 'setup', Object.keys(include('build/')), callback);
 });